Update launchpad output to be an array
The launchpad requires that the deposit data be nested in an array. So if you pipe the output to a file, the launchpad still says it's invalid, until you add `[ ]` to the output.

This will output as an array so it can be copied or piped to a file without any manual changes.

I can also see why you might not want to do this, so I'll leave it up to you if you want to merge or cancel this.